chore(clippy): root-crate cleanup batch 12 — long-tail singletons sweep (18 files)
Multi-category cleanup sweep targeting the long tail of remaining lints
(roughly 25 distinct lints across 18 files, each a 1-2 line mechanical
fix).

## `needless_question_mark` (Ok and ? unneeded)

- `src/network/arrow_ipc/codec.rs:189,201,364` — 3 sites collecting into
  Result<Vec<_>> and immediately re-wrapping in Ok(…?). Drop the wrap.

## `redundant_pattern_matching` / `useless_conversion`

- `src/storage/collection_pinning.rs:366`,
  `src/cluster/primary_pod_registry.rs:390` — `Error::new(ErrorKind::Other, e)`
  → `Error::other(e)` (stable since 1.74)

## `into_iter_on_owned`

- `src/network/arrow_ipc/service.rs:684`,
  `src/graph/query/executor.rs:145`,
  `src/storage/engines/raptor/consolidated_reader.rs:509` —
  `.into_iter()` call in IntoIterator-arg position is redundant

## `manual_clamp` → `.clamp(lo, hi)`

- `src/index/axis/hmgi/tiering.rs:312` — `((100.0-rate)/100.0).max(0.0).min(1.0)`
- `src/storage/engines/core/formats/proximablocks/spatial_clustering.rs:220,221`
  — `dimensions.min(64).max(1)` and `bits_per_dim.min(16).max(1)`

## `manual_repeat_n` (Rust 1.82+)

- `src/network/arrow_ipc/codec.rs:786` — `std::iter::repeat(x).take(n)`
  → `std::iter::repeat_n(x, n)`

## `needless_borrow` / `explicit_auto_deref`

- `src/network/arrow_ipc/service.rs:773` — `tenant_id.as_deref()`
  where `tenant_id: Option<&str>` (no-op) → `tenant_id`
- `src/storage/engines/swift/progressive_search.rs:738` — `&record`
  in for-loop where record is already `&ProximaRecord`
- `src/query/federated/optimizer/trace_replay.rs:90` — `&inputs`
  where `inputs: &ReplayInputs<'_>` already a ref

## `unnecessary_cast`

- `src/storage/engines/sst/flush/coordinator.rs:116` — `as u64` on
  Option<u64>::unwrap_or(0) (already u64)

## `manual_size_of_val`

- `src/storage/engines/sst/search/coordinator.rs:88` —
  `v.len() * size_of::<f32>()` → `size_of_val(v)`

## `match_like_matches_macro`

- `src/query/compute_bridge.rs:298` — multi-arm `match { Variant => true,
  _ => false }` → `matches!(…)`

## `iter_kv_map`

- `src/query/repair/sure_aggregator.rs:137` — `for (_claim_id, agg)
  in &per_claim` → `for agg in per_claim.values()`

## `manual_strip`

- `src/query/sql_frontend/parser.rs:3217` — `if input.starts_with('"')
  { let rest = &input[1..]; }` → `if let Some(rest) = input.strip_prefix('"')`

## `needless_not_is_some_and` → `is_none_or`

- `src/query/table_write_plan.rs:694` — `!primary_layout(t).is_some_and(
  |layout| !layout.lossy_type_mappings.is_empty())` → `is_none_or(|l|
  l.lossy_type_mappings.is_empty())`

## `manual_char_eq` → array pattern

- `src/query/multimodel_router.rs:147` — `matches!(ch, '\'' | … | ' ')`
  → `[' \'' , '"', ';', ',', ')', ' ']` (slice pattern accepted by
  `str::find`)

## `default_constructed_unit_structs`

- `src/services/operations/vectors/legacy.rs:1171` —
  `DefaultPseudoQueryGenerator::default()` → `DefaultPseudoQueryGenerator`
  (unit struct)

## Verification

- `cargo check -p proximadb --lib`: clean (13 unrelated dead-code warnings).
- All edits preserve semantics; nothing changed except clippy-pleasing
  syntactic form.

+++ src/storage/engines/sst/flush/coordinator.rs
@@ -113,7 +113,7 @@ impl FlushCoordinator {
             self.engine.tiering_integration(),
             params.collection_id.as_ref(),
         ) {
-            let bytes_written = result.bytes_written.unwrap_or(0) as u64;
+            let bytes_written = result.bytes_written.unwrap_or(0);
             tiering
                 .record_access(
                     coll.as_str(),
